Chat_With_SQL_DB/requirements.txt
streamlit
langchain
langchain_community
langchain_core
langchain_groq
langchain_cohere
sqlalchemy
python-dotenv